+++ src/Configuration/LibraryLoader.php
@@ -123,16 +123,10 @@ class LibraryLoader
 
             // try to load the DI configuration for the configured extension libraries
             foreach ($additionalVendorDir->getLibraries() as $library) {
-                if ($libraryDir = realpath(sprintf('%s/%s', $additionalVendorDir->getVendorDir(), $library))) {
+                if ($libraryDir = realpath(sprintf('%s/%s/%s', $configuration->getInstallationDir(), $additionalVendorDir->getVendorDir(), $library))) {
                     $this->loadConfiguration($customLoader, $magentoVersion, $libraryDir);
                 } else {
-                    throw new \Exception(
-                        sprintf(
-                            'Can\'t find find library directory "%s/%s"',
-                            $additionalVendorDir->getVendorDir(),
-                            $library
-                        )
-                    );
+                    throw new \Exception(sprintf('Can\'t find find library directory "%s"', $libraryDir));
                 }
             }
         }