+++ src/Command/ImportConvertValueCommand.php
@@ -24,7 +24,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Output\OutputInterface;
 use TechDivision\Import\Utils\CommandNames;
-use TechDivision\Import\ConfigurationInterface;
+use TechDivision\Import\Configuration\ConfigurationInterface;
 use TechDivision\Import\Serializers\ValueCsvSerializer;
 
 /**
@@ -62,9 +62,9 @@ class ImportConvertValueCommand extends AbstractSimpleImportCommand
     /**
      * Finally executes the simple command.
      *
-     * @param \TechDivision\Import\ConfigurationInterface       $configuration The configuration instance
-     * @param \Symfony\Component\Console\Input\InputInterface   $input         An InputInterface instance
-     * @param \Symfony\Component\Console\Output\OutputInterface $output        An OutputInterface instance
+     * @param \TechDivision\Import\Configuration\ConfigurationInterface $configuration The configuration instance
+     * @param \Symfony\Component\Console\Input\InputInterface           $input         An InputInterface instance
+     * @param \Symfony\Component\Console\Output\OutputInterface         $output        An OutputInterface instance
      *
      * @return void
      */