+++ src/Command/ImportCreateConfigurationFileCommand.php
@@ -30,7 +30,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use TechDivision\Import\Utils\CommandNames;
-use TechDivision\Import\ConfigurationInterface;
+use TechDivision\Import\Configuration\ConfigurationInterface;
 
 /**
  * The command implementation that creates a configuration file from one of the templates.
@@ -66,9 +66,9 @@ class ImportCreateConfigurationFileCommand extends AbstractSimpleImportCommand
     /**
      * Finally executes the simple command.
      *
-     * @param \TechDivision\Import\ConfigurationInterface       $configuration The configuration instance
-     * @param \Symfony\Component\Console\Input\InputInterface   $input         An InputInterface instance
-     * @param \Symfony\Component\Console\Output\OutputInterface $output        An OutputInterface instance
+     * @param \TechDivision\Import\Configuration\ConfigurationInterface $configuration The configuration instance
+     * @param \Symfony\Component\Console\Input\InputInterface           $input         An InputInterface instance
+     * @param \Symfony\Component\Console\Output\OutputInterface         $output        An OutputInterface instance
      *
      * @return void
      */